+++ client/src/actions/CardActions.js
@@ -1,6 +1,5 @@
 import apiClient from "../lib/ApiClient";
 import * as types from "../constants/ActionTypes";
-import { fetchBoard } from "./BoardActions";
 
 export function createCard(listId, title, callback) {
   return function (dispatch) {
@@ -61,19 +60,3 @@ export function editCard(obj, id, callback) {
   };
 }
 
-export function addCommentSuccess(data) {
-  return { type: types.ADD_COMMENT_SUCCESS, comment: data };
-}
-
-export function addComment(comment, cardId, callback) {
-  return function (dispatch) {
-    const obj = { cardId, comment: { text: comment } };
-    apiClient.addComment(obj, (data) => {
-      dispatch(addCommentSuccess(data));
-
-      if (callback) {
-        callback();
-      }
-    });
-  };
-}
